+++ error.php
@@ -22,6 +22,7 @@ $menu          = $app->getMenu()->getActive();
 $pageclass     = $menu !== null ? $menu->getParams()->get('pageclass_sfx', '') : '';
 $themeSwitcher = (boolean)$this->params->get('theme-switcher', 1);
 $fontAwesome   = (boolean)$this->params->get('font-awesome-thats-actually-rather-shit', 1);
+$googleFont    = $this->params->get('google-font', '');
 
 // Template params
 if ($themeSwitcher)
@@ -36,6 +37,19 @@ $wa->usePreset('template.lightning');
 if ($fontAwesome)
 {
 	$wa->useStyle('fontawesome.css');
+}
+
+// Google font
+if ($googleFont !== '')
+{
+	$fontFamily = str_replace(' ', '+', $googleFont);
+	$this->getPreloadManager()->preconnect('https://fonts.googleapis.com', ['crossorigin' => 'anonymous']);
+	$this->getPreloadManager()->preconnect('https://fonts.gstatic.com', ['crossorigin' => 'anonymous']);
+	$this->addHeadLink('https://fonts.googleapis.com/css2?family=' . $fontFamily . '&display=swap', 'stylesheet', 'rel');
+
+	$wa->addInlineStyle(':root {
+		--hiq-font-family-base: "' . $googleFont . '";
+	}');
 }
 
 // Load switcher JS