+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## WIP
+- Fix loading Google Fonts on error and offline pages
+
 ## 1.2.0
 - Add support for Google fonts
 - Add text colour utility classes